make sure classic-rest working in native
