+++ runtime/src/main/resources/META-INF/services/jakarta.ws.rs.ext.Providers
@@ -1,0 +1,1 @@
+io.quarkiverse.fury.ClientFurySerializer
